pyproject.toml
[project]
name = "freight-ops"
version = "0.1.0"
description = "AI-powered freight operations platform for hotshot trucking"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["freight", "trucking", "ai-agents", "logistics", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    # Core AI/LLM
    "anthropic>=0.40.0",
    "openai>=1.54.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",

    # MCP Protocol
    "mcp>=1.1.0",
    "open-ptc-agent>=0.1.0",

    # Data validation and models
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # API clients and HTTP
    "httpx>=0.27.0",
    "aiohttp>=3.11.0",
    "requests>=2.32.0",

    # Data processing
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "polars>=1.12.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",

    # Configuration
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "dynaconf>=3.2.0",

    # Date/time handling
    "python-dateutil>=2.9.0",
    "arrow>=1.3.0",

    # Geospatial and routing
    "geopy>=2.4.0",
    "googlemaps>=4.10.0",
    "haversine>=2.8.0",

    # Financial calculations
    "decimal>=1.0.0",
    "money>=1.3.0",

    # Caching
    "redis>=5.2.0",
    "diskcache>=5.6.0",

    # Logging and monitoring
    "structlog>=24.4.0",
    "loguru>=0.7.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.9.0",

    # Async utilities
    "tenacity>=9.0.0",
    "asyncio-throttle>=1.0.0",

    # JSON processing
    "orjson>=3.10.0",

    # Environment detection
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.0",
    "hypothesis>=6.115.0",

    # Linting and formatting
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "black>=24.10.0",

    # Type stubs
    "types-requests>=2.32.0",
    "types-PyYAML>=6.0.0",
    "types-python-dateutil>=2.9.0",
    "pandas-stubs>=2.2.0",

    # Development tools
    "ipython>=8.29.0",
    "ipdb>=0.13.0",

    # Documentation
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]

notebooks = [
    "jupyter>=1.1.0",
    "jupyterlab>=4.3.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "plotly>=5.24.0",
]

mcp-servers = [
    # Dependencies for custom MCP servers
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=13.1.0",
]

all = [
    "freight-ops[dev,notebooks,mcp-servers]",
]

[project.scripts]
freight-ops = "src.cli:app"
freight-dispatch = "src.agents.dispatch:main"
freight-rates = "src.agents.rate_analysis:main"
freight-compliance = "src.agents.compliance:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "freight-ops[dev]",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG", "S101"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "api: Tests that call external APIs",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "open_ptc_agent.*",
    "googlemaps.*",
    "geopy.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
